+++ app/code/Svea/Maksuturva/Model/Gateway/Base.php
@@ -64,6 +64,7 @@ abstract class Base extends \Magento\Framework\Model\AbstractModel
      */
     protected $config;
 
+    protected $_loggerHandler;
 
     /**
      * Base constructor.
@@ -98,6 +99,22 @@ abstract class Base extends \Magento\Framework\Model\AbstractModel
         } else {
             throw new \Svea\Maksuturva\Model\Gateway\Exception(array('the hash algorithms SHA-512, SHA-256, SHA-1 and MD5 are not supported!'), self::EXCEPTION_CODE_ALGORITHMS_NOT_SUPORTED);
         }
+    }
+
+    public function sveaLoggerInfo($info)
+    {
+        $this->getSveaLoggerHandler()->info(json_encode($info));
+    }
+
+    protected function getSveaLoggerHandler()
+    {
+        if(!$this->_loggerHandler) {
+            $writer = new \Zend_Log_Writer_Stream(BP . '/var/log/svea-payment-module.log');
+            $logger = new \Zend_Log();
+            $logger->addWriter($writer);
+            $this->_loggerHandler = $logger;
+        }
+        return $this->_loggerHandler;
     }
 
     public function generateReturnHash($hashData)
@@ -212,6 +229,8 @@ abstract class Base extends \Magento\Framework\Model\AbstractModel
                 );
             }
         }
+
+        $this->sveaLoggerInfo("Cancel or refund result for pmtc_id" . $parsedResponse['pmtc_id'] . " is '" . $parsedResponse['pmtc_returntext'] . "'" );
 
         switch($parsedResponse['pmtc_returncode']){